added mathes and so on, working stuff

+++ client/src/pages/Team.js
@@ -3,6 +3,9 @@ import axios from 'axios';
 import Table from 'react-bootstrap/Table';
 import TeamModal from '../components/TeamModal'; // Предполагается, что вы создали этот компонент
 import EditTeamModal from '../components/EditTeamModal'; // Предполагается, что вы создали этот компонент
+// import Button from 'react-bootstrap/Button';
+import { toast } from 'react-toastify';
+
 
 function Team () {
 	const [teams, setTeams] = useState([]);
@@ -15,42 +18,54 @@ function Team () {
 	const openEditModal = () => setShowEditModal(true);
 	const closeEditModal = () => setShowEditModal(false);
 
-	// const createTeam = async (formData) => {
-	// 	fetch('http://localhost:5000/api/team', {
-	// 		method: 'POST',
-	// 		body: formData
-	// 	}).then(res => {
-	// 		console.log(res);
-	// 		setTeams([...teams, res]);
-	// 		closeModal();
-	// 	})
-	// 		.catch(err => {
-	// 			console.error(`Error: ${err}`);
-	// 		})
-	// };
-
-
 	const deleteTeam = (id) => {
-		axios.delete(`http://localhost:5000/api/team/teamDel/${id}`)
+		axios.get(`http://localhost:5000/api/match`)
 			.then(response => {
-				setTeams(teams.filter(team => team.team_id !== id));
+				let isExist = false;
+				response.data.forEach(element => {
+					if (element.team1_id === id || element.team2_id === id) {
+						isExist = true
+					}
+				});
+				if (isExist) {
+					toast.error('Cannot delete team with associated matches');
+				} else {
+					console.log('EREOROEROEORE');
+					axios.delete(`http://localhost:5000/api/team/teamDel/${id}`)
+						.then(response => {
+							setTeams(teams.filter(team => team.team_id !== id));
+						})
+						.catch(error => {
+							console.error(`Error: ${error}`);
+							toast.error('Error deleting team');
+						});
+				}
 			})
 			.catch(error => {
 				console.error(`Error: ${error}`);
+				toast.error('Error fetching matches for team');
 			});
 	};
 
-	const updateTeam = (id, updatedTeamData) => {
-		axios.put(`http://localhost:5000/api/team/teamEdit/${id}`, updatedTeamData)
-			.then(response => {
-				setTeams(teams.map(team =>
-					team.team_id === id ? { ...team, ...updatedTeamData } : team
-				));
-				closeEditModal();
-			})
-			.catch(error => {
-				console.error(`Error: ${error}`);
-			});
+
+	const updateTeam = async (teamData) => {
+		const response = await fetch(`http://localhost:5000/api/team/teamEdit/${editingTeam.team_id}`, {
+			method: 'PUT',
+			body: teamData,
+		});
+
+		if (!response.ok) {
+			throw new Error('Network response was not ok');
+		}
+
+		const updatedTeamResponse = await fetch(`http://localhost:5000/api/team/${editingTeam.team_id}`);
+		const updatedTeam = await updatedTeamResponse.json();
+
+		setTeams(teams.map(team =>
+			team.team_id === editingTeam.team_id ? updatedTeam : team
+		));
+
+		closeEditModal();
 	};
 
 	useEffect(() => {
@@ -69,6 +84,7 @@ function Team () {
 			<Table striped bordered hover>
 				<thead>
 					<tr>
+						<th>Team ID</th>
 						<th>Image</th>
 						<th>Team Name</th>
 						<th>Country</th>
@@ -81,7 +97,10 @@ function Team () {
 				<tbody>
 					{teams.map(team => (
 						<tr key={team.team_id}>
-							<td><img src={`http://localhost:5000/${team.img}`} alt={team.team_name} style={{ width: '70px', height: '70px' }} /></td>
+							<td>{team.team_id}</td>
+							<td>
+								{team.img && <img src={`http://localhost:5000/${team.img}`} alt={team.team_name} style={{ width: '70px', height: '70px' }} />}
+							</td>
 							<td>{team.team_name}</td>
 							<td>{team.team_country}</td>
 							<td>{team.date_of_creating_team}</td>
@@ -116,7 +135,7 @@ function Team () {
 			<EditTeamModal
 				show={showEditModal}
 				onClose={closeEditModal}
-				onUpdate={(updatedTeamData) => updateTeam(editingTeam.team_id, updatedTeamData)}
+				onUpdate={updateTeam}
 				editingTeam={editingTeam}
 			/>
